Stop trying to put static files in the binary
