+++ build.sh
@@ -4,6 +4,7 @@ set -e
 
 version=$(date +%Y%m%d%H%M)
 
+cd $(dirname $0)
 for img in moth moth-devel; do
     echo "==== $img"
     sudo docker build --build-arg http_proxy=$http_proxy --build-arg https_proxy=$https_proxy --tag dirtbags/$img --tag dirtbags/$img:$version -f Dockerfile.$img .
